+++ src/app/config/setup/personalization/personalization.service.ts
@@ -3,7 +3,7 @@ import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { defaultIfEmpty, map } from 'rxjs/operators';
 
-import { OctoprintPrinterProfiles } from '../../../model/octoprint/printer-profile.model';
+import { OctoprintPrinterProfiles } from '../../../model/octoprint';
 
 @Injectable({
   providedIn: 'root',
