+++ src/app/control/custom-actions/custom-actions.component.ts
@@ -1,9 +1,9 @@
 import { Component, Input } from '@angular/core';
 import { SafeResourceUrl } from '@angular/platform-browser';
 import { Router } from '@angular/router';
-import { PSUState } from 'src/app/model';
 
 import { ConfigService } from '../../config/config.service';
+import { PSUState } from '../../model';
 import { EnclosureService } from '../../services/enclosure/enclosure.service';
 import { PrinterService } from '../../services/printer/printer.service';
 import { SystemService } from '../../services/system/system.service';
