+++ src/app/model/index.ts
@@ -1,7 +1,9 @@
 export * from './auth.model';
 export * from './enclosure.model';
+export * from './event.model';
 export * from './filament.model';
 export * from './files.model';
 export * from './job.model';
 export * from './printer-profile.model';
 export * from './printer.model';
+export * from './system.model';
