+++ src/app/model/octoprint/connection.model.ts
@@ -1,25 +1,3 @@
-export interface OctoprintConnection {
-  current: OctoprintCurrentConnection;
-  options: OctoprintConnectionOptions;
-}
-
-interface OctoprintCurrentConnection {
-  state: string;
-  port: string;
-  baudrate: number;
-  printerProfile: string;
-}
-
-interface OctoprintConnectionOptions {
-  ports: string[];
-  baudrates: number[];
-  printerProfiles: Record<string, unknown>;
-  portPreference: string;
-  baudratePreference: string;
-  printerProfilePreference: string;
-  autoconnect: boolean;
-}
-
 export interface ConnectCommand {
   command: string;
   port?: string;
