+++ src/app/services/socket/socket.service.ts
@@ -1,8 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
-import { PrinterEvent } from 'src/app/model/event.model';
 
-import { JobStatus, PrinterStatus } from '../../model';
+import { JobStatus, PrinterEvent, PrinterStatus } from '../../model';
 
 @Injectable()
 export abstract class SocketService {
